+++ src/components/patterns/index.js
@@ -1,7 +1,9 @@
 import Grid from './Grid'
 import Footer from './Footer'
 import Toolbar from './Toolbar'
+import LiveEditor from './LiveEditor'
+import CardLink from './CardLink'
 import PricingTable from './PricingTable'
 import List from './List'
 
-export { Grid, List, Footer, Toolbar, PricingTable }
+export { CardLink, LiveEditor, Grid, List, Footer, Toolbar, PricingTable }
