+++ src/pages/index.js
@@ -1,21 +1,15 @@
-import is from 'styled-is'
 import React, { Fragment, Component } from 'react'
-import { transition, space } from 'theme'
-import styled from 'styled-components'
-import { BlockLink } from 'rebass'
 import {
   Text,
   Subhead,
   Box,
-  Caps,
   Heading,
   Lead,
-  Card,
   Flex,
   Container,
   Hide
 } from 'components/elements'
-import { PricingTable, Grid } from 'components/patterns'
+import { CardLink, LiveEditor, PricingTable, Grid } from 'components/patterns'
 import { List, ListItem } from 'components/patterns/List'
 import {
   Working,
@@ -23,80 +17,6 @@ import {
   DesignProcess,
   Frameworks
 } from 'components/icons'
-
-const CapsIcon = styled(Caps)`
-  transition: margin-left ${transition.medium};
-
-  ${is('hover')`
-  margin-left: ${space[2]}px;
-`};
-`
-
-CapsIcon.defaultProps = {
-  blacklist: [...Object.keys(Caps.propTypes), 'hover']
-}
-
-const CardTitle = ({ children, hover }) => (
-  <Text my={27}>
-    <Caps
-      is='span'
-      fontWeight='bold'
-      color='secondary'
-      fontSize={2}
-      children={children}
-    />
-    <CapsIcon
-      is='span'
-      fontWeight='bold'
-      color='secondary'
-      fontSize={2}
-      ml={1}
-      children='→'
-      hover={hover}
-    />
-  </Text>
-)
-
-const CardLink = class extends Component {
-  state = { hover: false }
-  mouseOut = () => this.setState({ hover: false })
-  onMouseOver = () => this.setState({ hover: true })
-  render () {
-    const {
-      iconComponent: IconComponent,
-      title,
-      description,
-      href
-    } = this.props
-    const { hover } = this.state
-    return (
-      <BlockLink
-        href={href}
-        onMouseOut={this.mouseOut}
-        onMouseOver={this.onMouseOver}
-      >
-        <Card py={[47.6, 56]} px={4} width={[314.5, 370]} height={[400, 420]}>
-          <Flex
-            justifyContent='space-between'
-            alignItems='center'
-            flexDirection='column'
-            style={{ height: '100%' }}
-          >
-            <IconComponent width='100%' />
-            <CardTitle children={title} hover={hover} />
-            <Text
-              fontSize={15}
-              textAlign='center'
-              color='black80'
-              lineHeight={3}
-              children={description}
-            />
-          </Flex>
-        </Card>
-      </BlockLink>
-    )
-  }
-}
 
 export default class extends Component {
   render () {
@@ -120,6 +40,11 @@ export default class extends Component {
               />
               <Lead color='black50' children='Enter an URL, receive data.' />
             </Flex>
+            <Box is='article'>
+              <Container is='section' py={[2, 3]}>
+                <LiveEditor />
+              </Container>
+            </Box>
           </Container>
         </Box>
         <Box bg='#F5F4F9' is='article'>
